+++ plugins/clipboard/plugin.js
@@ -1028,8 +1028,7 @@
 					type: 'auto',
 					method: 'paste',
 					dataTransfer: clipboard.initPasteDataTransfer( evt )
-				},
-				external = eventData.dataTransfer.getTransferType( editor ) === CKEDITOR.DATA_TRANSFER_EXTERNAL;
+				};
 
 			eventData.dataTransfer.cacheData();
 
@@ -1041,7 +1040,7 @@
 			var beforePasteNotCanceled = editor.fire( 'beforePaste', eventData ) !== false;
 
 			// Do not use paste bin if the browser let us get HTML or files from dataTranfer.
-			if ( beforePasteNotCanceled && ( clipboard.isHtmlInExternalDataTransfer || !external ) && !eventData.dataTransfer.isEmpty() ) {
+			if ( beforePasteNotCanceled && clipboard.canClipboardApiBeTrusted( eventData.dataTransfer, editor ) ) {
 				evt.data.preventDefault();
 				setTimeout( function() {
 					firePasteEvents( editor, eventData );
@@ -1489,17 +1488,6 @@
 		 */
 		isFileApiSupported: !CKEDITOR.env.ie || CKEDITOR.env.version > 9,
 
-
-		/**
-		 * True if external data transfer contains HTML data.
-		 *
-		 * @since 4.5
-		 * @readonly
-		 * @property {Boolean}
-		 */
-		isHtmlInExternalDataTransfer: !CKEDITOR.env.ie && !CKEDITOR.env.safari,
-
-
 		/**
 		 * Main native paste event editable should listen to.
 		 *
@@ -1514,6 +1502,40 @@
 		 * @property {String}
 		 */
 		mainPasteEvent: ( CKEDITOR.env.ie && !CKEDITOR.env.edge ) ? 'beforepaste' : 'paste',
+
+		/**
+		 * Returns `true` if we can expect that the browser provides HTML data through the Clipboard API.
+		 * If not, then it returns `false` and as a result CKEditor will use the paste bin. Read more in
+		 * the [Clipboard Integration](http://docs.ckeditor.com/#!/guide/dev_clipboard-section-clipboard-api) guide.
+		 *
+		 * @since 4.5.2
+		 * @returns {Boolean}
+		 */
+		canClipboardApiBeTrusted: function( dataTransfer, editor ) {
+			// If it's an internal or cross-editor data transfer, then it means that custom cut/copy/paste support works
+			// and that the data were put manually on the data transfer so we can be sure that it's available.
+			if ( dataTransfer.getTransferType( editor ) != CKEDITOR.DATA_TRANSFER_EXTERNAL ) {
+				return true;
+			}
+
+			// In Chrome we can trust Clipboard API, with the exception of Chrome on Android (in both - mobile and desktop modes), where
+			// clipboard API is not available so we need to check it (#13187).
+			if ( CKEDITOR.env.chrome && !dataTransfer.isEmpty() ) {
+				return true;
+			}
+
+			// Because of a Firefox bug HTML data are not available in some cases (e.g. paste from Word), in such cases we
+			// need to use the pastebin (#13528, https://bugzilla.mozilla.org/show_bug.cgi?id=1183686).
+			if ( CKEDITOR.env.gecko && ( dataTransfer.getData( 'text/html' ) || dataTransfer.getFilesCount() ) ) {
+				return true;
+			}
+
+			// In Safari and IE HTML data is not available though the Clipboard API.
+			// In Edge things are a bit messy at the moment -
+			// https://connect.microsoft.com/IE/feedback/details/1572456/edge-clipboard-api-text-html-content-messed-up-in-event-clipboarddata
+			// It is safer to use the paste bin in unknown cases.
+			return false;
+		},
 
 		/**
 		 * Returns the element that should be used as the target for the drop event.